# server/swaggerDoc.json
{
  "swagger": "2.0",
  "info": {
    "description": " This is the api documentation for the Monash Tutor Allocation System",
    "version": "1.0.0",
    "title": "Monash Tutor Allocation System"
  },
...
  "schemes": ["http"],
  "paths": {
...
    "/allocations": {
      "post": {
        "tags": ["Allocation"],
        "summary": "Creates an allocation",
        "description": "\n      \n      Role authorisation:\n      - TA: not allowed\n      - Lecturer: can create allocation for unit they're lecturing but not isWorkforceApproved attribute\n      - Admin: can create allocation in any unit\n      ",
        "operationId": "createAllocation",
        "consumes": ["application.json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "New allocation object",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "id": { "type": "string", "format": "uuid" },
                "activityId": { "type": "string", "format": "uuid" },
                "staffId": { "type": "string", "format": "uuid" },
                "offerExpiryDate": {
                  "type": "string",
                  "format": "timestamptz"
                },
                "isLecturerApproved": {
                  "type": "boolean",
                  "default": null,
                  "nullable": true
                },
                "isTaAccepted": {
                  "type": "boolean",
                  "default": null,
                  "nullable": true
                },
                "isWorkforceApproved": {
                  "type": "boolean",
                  "default": null,
                  "nullable": true
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "properties": {
                "id": { "type": "string", "format": "uuid" },
                "activityId": { "type": "string", "format": "uuid" },
                "staffId": { "type": "string", "format": "uuid" },
                "offerExpiryDate": {
                  "type": "string",
                  "format": "timestamptz"
                },
                "isLecturerApproved": {
                  "type": "boolean",
                  "default": null,
                  "nullable": true
                },
                "isTaAccepted": {
                  "type": "boolean",
                  "default": null,
                  "nullable": true
                },
                "isWorkforceApproved": {
                  "type": "boolean",
                  "default": null,
                  "nullable": true
                }
              }
            }
          },
          "401": {
            "description": "user has not been authenticated",
            "schema": {
              "type": "object",
              "properties": {
                "authenticated": { "type": "boolean", "example": "false" },
                "message": {
                  "type": "string",
                  "example": "user has not been authenticated"
                }
              }
            }
          },
          "403": {
            "description": "User role doesn't have authorisation for the operation",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "UnauthorisedAccessedError"
                },
                "message": { "type": "string" }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Allocation"],
        "summary": "Updates an allocation",
        "description": "\n      \n      Role authorisation:\n      - TA: not allowed\n      - Lecturer: not allowed\n      - Admin: can update allocation in any unit, regardless of the acceptance status\n      ",
        "operationId": "updateAllocation",
        "consumes": ["application.json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Updated allocation object",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "id": { "type": "string", "format": "uuid" },
                "activityId": { "type": "string", "format": "uuid" },
                "staffId": { "type": "string", "format": "uuid" },
                "offerExpiryDate": {
                  "type": "string",
                  "format": "timestamptz"
                },
                "isLecturerApproved": {
                  "type": "boolean",
                  "default": null,
                  "nullable": true
                },
                "isTaAccepted": {
                  "type": "boolean",
                  "default": null,
                  "nullable": true
                },
                "isWorkforceApproved": {
                  "type": "boolean",
                  "default": null,
                  "nullable": true
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "properties": {
                "id": { "type": "string", "format": "uuid" },
                "activityId": { "type": "string", "format": "uuid" },
                "staffId": { "type": "string", "format": "uuid" },
                "offerExpiryDate": {
                  "type": "string",
                  "format": "timestamptz"
                },
                "isLecturerApproved": {
                  "type": "boolean",
                  "default": null,
                  "nullable": true
                },
                "isTaAccepted": {
                  "type": "boolean",
                  "default": null,
                  "nullable": true
                },
                "isWorkforceApproved": {
                  "type": "boolean",
                  "default": null,
                  "nullable": true
                }
              }
            }
          },
          "401": {
            "description": "user has not been authenticated",
            "schema": {
              "type": "object",
              "properties": {
                "authenticated": { "type": "boolean", "example": "false" },
                "message": {
                  "type": "string",
                  "example": "user has not been authenticated"
                }
              }
            }
          },
          "403": {
            "description": "User role doesn't have authorisation for the operation",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "UnauthorisedAccessedError"
                },
                "message": { "type": "string" }
              }
            }
          }
        }
      }
    },
    "/allocations/mine": {
      "get": {
        "tags": ["Allocation"],
        "summary": "Get the allocated activities of the current user",
        "operationId": "getMyAllocation",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "query",
            "name": "unitId",
            "description": "Filter allocations with unitId",
            "required": false,
            "type": "string"
          },
          {
            "in": "query",
            "name": "isLecturerApproved",
            "description": "Filter allocations with lecturerApproved true or false",
            "required": false,
            "type": "boolean"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": { "type": "string", "format": "uuid" },
                  "activityId": { "type": "string", "format": "uuid" },
                  "staffId": { "type": "string", "format": "uuid" },
                  "offerExpiryDate": {
                    "type": "string",
                    "format": "timestamptz"
                  },
                  "isLecturerApproved": {
                    "type": "boolean",
                    "default": null,
                    "nullable": true
                  },
                  "isTaAccepted": {
                    "type": "boolean",
                    "default": null,
                    "nullable": true
                  },
                  "isWorkforceApproved": {
                    "type": "boolean",
                    "default": null,
                    "nullable": true
                  },
                  "staff": {
                    "type": "object",
                    "properties": {
                      "id": { "type": "string", "format": "uuid" },
                      "givenNames": { "type": "string", "nullable": true },
                      "lastName": { "type": "string", "nullable": true },
                      "aqf": { "type": "integer", "nullable": true },
                      "studyingAqf": { "type": "integer", "nullable": true },
                      "email": {
                        "uniqueItems": true,
                        "type": "string",
                        "format": "text"
                      },
                      "googleId": { "type": "string", "nullable": true },
                      "appRole": { "type": "string", "enum": ["User", "Admin"] }
                    }
                  },
                  "unit": {
                    "type": "object",
                    "properties": {
                      "id": { "type": "string", "format": "uuid" },
                      "unitCode": { "type": "string" },
                      "offeringPeriod": { "type": "string" },
                      "campus": { "type": "string" },
                      "year": { "type": "integer" },
                      "aqfTarget": { "type": "integer" }
                    }
                  },
                  "activity": {
                    "type": "object",
                    "properties": {
                      "id": { "type": "string", "format": "uuid" },
                      "activityCode": { "type": "string" },
                      "activityGroup": { "type": "string" },
                      "campus": { "type": "string" },
                      "location": { "type": "string" },
                      "dayOfWeek": {
                        "type": "string",
                        "enum": ["M", "T", "W", "Th", "F"]
                      },
                      "startTime": { "type": "string", "format": "time" },
                      "endTime": { "type": "string", "format": "time" },
                      "unitId": { "type": "string", "format": "uuid" },
                      "studentCount": { "type": "integer", "default": 0 },
                      "allocationMaxNum": { "type": "integer", "default": 1 }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "user has not been authenticated",
            "schema": {
              "type": "object",
              "properties": {
                "authenticated": { "type": "boolean", "example": "false" },
                "message": {
                  "type": "string",
                  "example": "user has not been authenticated"
                }
              }
            }
          }
        }
      }
    },
    "/allocations/unswapped": {
      "get": {
        "tags": ["Allocation"],
        "summary": "Get unswapped allocations",
        "operationId": "getUnswappedAllocation",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "query",
            "name": "unitId",
            "description": "Filter allocations with unitId",
            "required": false,
            "type": "string"
          },
          {
            "in": "query",
            "name": "isLecturerApproved",
            "description": "Filter allocations with isLecturerApproved true or false",
            "required": false,
            "type": "boolean"
          },
          {
            "in": "query",
            "name": "isWorkforceApproved",
            "description": "Filter allocations with isWorkforceApproved true or false",
            "required": false,
            "type": "boolean"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": { "type": "string", "format": "uuid" },
                  "activityId": { "type": "string", "format": "uuid" },
                  "staffId": { "type": "string", "format": "uuid" },
                  "offerExpiryDate": {
                    "type": "string",
                    "format": "timestamptz"
                  },
                  "isLecturerApproved": {
                    "type": "boolean",
                    "default": null,
                    "nullable": true
                  },
                  "isTaAccepted": {
                    "type": "boolean",
                    "default": null,
                    "nullable": true
                  },
                  "isWorkforceApproved": {
                    "type": "boolean",
                    "default": null,
                    "nullable": true
                  },
                  "staff": {
                    "type": "object",
                    "properties": {
                      "id": { "type": "string", "format": "uuid" },
                      "givenNames": { "type": "string", "nullable": true },
                      "lastName": { "type": "string", "nullable": true },
                      "aqf": { "type": "integer", "nullable": true },
                      "studyingAqf": { "type": "integer", "nullable": true },
                      "email": {
                        "uniqueItems": true,
                        "type": "string",
                        "format": "text"
                      },
                      "googleId": { "type": "string", "nullable": true },
                      "appRole": { "type": "string", "enum": ["User", "Admin"] }
                    }
                  },
                  "unit": {
                    "type": "object",
                    "properties": {
                      "id": { "type": "string", "format": "uuid" },
                      "unitCode": { "type": "string" },
                      "offeringPeriod": { "type": "string" },
                      "campus": { "type": "string" },
                      "year": { "type": "integer" },
                      "aqfTarget": { "type": "integer" }
                    }
                  },
                  "activity": {
                    "type": "object",
                    "properties": {
                      "id": { "type": "string", "format": "uuid" },
                      "activityCode": { "type": "string" },
                      "activityGroup": { "type": "string" },
                      "campus": { "type": "string" },
                      "location": { "type": "string" },
                      "dayOfWeek": {
                        "type": "string",
                        "enum": ["M", "T", "W", "Th", "F"]
                      },
                      "startTime": { "type": "string", "format": "time" },
                      "endTime": { "type": "string", "format": "time" },
                      "unitId": { "type": "string", "format": "uuid" },
                      "studentCount": { "type": "integer", "default": 0 },
                      "allocationMaxNum": { "type": "integer", "default": 1 }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "user has not been authenticated",
            "schema": {
              "type": "object",
              "properties": {
                "authenticated": { "type": "boolean", "example": "false" },
                "message": {
                  "type": "string",
                  "example": "user has not been authenticated"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}